updated notification message in constant
